finish other ruleset logic
