+++ src/module/hooks/index.ts
@@ -1,4 +1,4 @@
-//NOTE! *Every* hook must be imported here, or they won't be used!
+// NOTE! *Every* hook must be imported here, or they won't be used!
 import './deleteItem';
 import './hooks';
 import './preCreateActor';
@@ -10,3 +10,4 @@ import './renderSettingsConfig';
 import './setup';
 import './updateHits';
 import './updateItem';
+import './showWoundIcons';
